feat(SPV-1026): no P2PStatus
