+++ engine/definitions.go
@@ -83,7 +83,6 @@ const (
 	metadataField        = "metadata"
 	nextExternalNumField = "next_external_num"
 	nextInternalNumField = "next_internal_num"
-	p2pStatusField       = "p2p_status"
 	satoshisField        = "satoshis"
 	spendingTxIDField    = "spending_tx_id"
 	statusField          = "status"
